+++ src/index.ts
@@ -14,7 +14,6 @@ import {
   deploy as deployInternal,
   getAccountProvider,
   getAddress as getAddressInternal,
-  getInitCode,
   isDeployed as isDeployedInternal,
   OwnersFieldRequiredError,
   setup as setupInternal,
@@ -568,7 +567,6 @@ export {
   RhinestoneSDK,
   walletClientToAccount,
   wrapParaAccount,
-  getInitCode,
   // Registry functions
   getSupportedTokens,
   getTokenAddress,